wealthsimple-operator/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
google-genai>=1.5.0
httpx>=0.24.0
